requirements.txt
albumentations==1.3.0
colorama==0.4.6
coremltools==6.3.0
Flask==2.3.2
ipython==8.14.0
matplotlib==3.7.1
mlsocket==0.1.2
numpy==1.24.3
onnx==1.14.0
onnxruntime_gpu==1.15.0
onnxsim==0.4.31
open3d==0.17.0
opencv_python==4.7.0.72
openvino==2023.0.0
pafy==0.5.5
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pycocotools==2.0.6
PyYAML==6.0
PyYAML==6.0
Random_Word==1.0.11
Requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
setuptools==67.8.0
tensorrt==8.6.1
tflite_runtime==2.12.0
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
